# ENH: sparse: speed up CSR/CSC toarray()