This PR adds a new function to sparsetools: `csr_todense`, which has the same semantics as `coo_todense` in that it's actually performing `D += S`. Unlike the COO version, this function requires that the dense array is in C-contiguous order.

@pv: I've never added functions to sparsetools before, so please let me know if I've missed something.

The one backwards-incompatible change here is that I set the default `csc_matrix.toarray` order to be F-contiguous, for efficiency reasons. Users can still specify an order explicitly, but the `mat.A` shorthand will now produce column-major matrices for CSC inputs.

There weren't any benchmarks testing `toarray` timing, so I added one. Here are the changes between current master and this PR:
```
BENCHMARK                                    BEFORE      AFTER     FACTOR
sparse.Densify.time_toarray('csc', 'F')    635.97μs   474.44μs   0.74600847x
sparse.Densify.time_toarray('csr', 'C')    629.63μs   468.08μs   0.74341441x
```

Finally, once gh-7079 is merged, this change will allow a more efficient implementation of `_add_dense`.