@pv 2017-03-12
LGTM, although I would leave the backward compat break out.