@perimosocordiae 2017-03-14
I rewrote some history and split this up a bit.

CSR/CSC + dense now avoids conversion to COO, enabled by gh-7079.

The backwards-incompatibility is now isolated to commit 726239eb0fe71ebf10deb91bee753b4138f11f9a. My personal opinion is that this won't adversely affect much user code, and that the upcoming 1.0 release seems like as good a time as we'll get for changes like this. Also, if the previous behavior is needed, users can always specify an explicit order in `.toarray()`.

That said, I'll defer to the consensus on whether it's worth breaking compatibility for the speedup here.